Keep a copy of modified system environment variables
It is important that we keep a copy of the modified system environment
variables. We modify these environment variables to establish a sandbox
for our app image. On the other hand, these changes could cause
conflicting libraries when calling external apps from within the app
image. Therefore it is important that the calling software restore the
system environment before executing external software via QProcess, etc.

Signed-off-by: Ömer Göktaş <omer@omergoktas.com>